Move `InputWidget` styles
