+++ src/styles.ts
@@ -1,5 +1,7 @@
 import { I18nManager, StyleSheet } from 'react-native';
 
+const isRtl = I18nManager.isRTL;
+
 export const styles = StyleSheet.create({
   panel_container: {
     position: 'relative',
@@ -16,7 +18,7 @@ export const styles = StyleSheet.create({
   },
   handle: {
     position: 'absolute',
-    ...(I18nManager.isRTL ? { right: 0 } : { left: 0 }),
+    ...(isRtl ? { right: 0 } : { left: 0 }),
     top: 0,
     justifyContent: 'center',
     alignItems: 'center',
@@ -35,5 +37,40 @@ export const styles = StyleSheet.create({
     shadowRadius: 3.84,
 
     elevation: 5,
+  },
+
+  // Input Widgets style
+  container: {
+    flexDirection: 'row',
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  arrowButton: {
+    width: 24,
+    height: 24,
+    marginLeft: 10,
+  },
+  inputsWrapper: {
+    flex: 1,
+    flexDirection: isRtl ? 'row-reverse' : 'row',
+    gap: 5,
+  },
+  inputsContainer: {
+    flex: 1,
+  },
+  input: {
+    borderWidth: 1,
+    borderColor: 'gray',
+    borderRadius: 4,
+    paddingVertical: 6,
+    paddingHorizontal: 10,
+    textAlign: 'center',
+    fontSize: 16,
+    color: '#000',
+  },
+  inputTitle: {
+    textAlign: 'center',
+    color: 'gray',
+    paddingVertical: 5,
   },
 });
